+++ ebzl/modules/list.py
@@ -10,7 +10,7 @@ def get_argument_parser():
     parser = argparse.ArgumentParser("ebzl list")
 
     parameters.add_profile(parser)
-    parameters.add_app_name(parser)
+    parameters.add_app_name(parser, required=False)
     parameters.add_region(parser, required=False)
 
     return parser